Cypress: refactor for Redact and NoRedact tests (#656)
