+++ acceptanceTests/cypress/integration/testHelpers/StatusBarHelper.js
